feat: topology editor UI with WASM validation (phases 1-8)
Interactive React topology editor for NebulaStream with:
- Drag-and-drop canvas for sources, sinks, workers, queries (ReactFlow)
- YAML serialization/deserialization with bidirectional sync
- Monaco SQL editor with ANTLR-based validation and completions
- Real-time WASM validation using nes-validator compiled via Emscripten
- Source/sink config validation via extracted CMake targets
- Playwright E2E and Vitest unit test suites
- Dark mode, keyboard shortcuts, history support

+++ nes-common/src/Util/UUID.cpp
@@ -17,6 +17,7 @@
 #include <array>
 #include <bit>
 #include <cstddef>
+#include <cstdint>
 #include <optional>
 #include <string>
 #include <utility>
@@ -60,6 +61,6 @@ NES::UUID NES::stringToUUIDOrThrow(const std::string& uuidString)
 
 size_t std::hash<NES::UUID>::operator()(const NES::UUID& uuid) const noexcept
 {
-    const auto [right, left] = std::bit_cast<std::array<size_t, 2>>(uuid);
+    const auto [right, left] = std::bit_cast<std::array<uint64_t, 2>>(uuid);
     return folly::hash::hash_combine(right, left);
 }
